Demos/Stock Market Chat/requirements.txt
quart>=3.0.0
agno>=2.2.1
openai>=1.109.1
python-dotenv>=1.0.0
mcp>=1.0.0
tenacity>=9.0.0
